chore(serials): split schemas.ts into per-domain files
schemas.ts (366 lines) was over the repo's 330-line max-lines lint
limit. Split into schemas/search, archive, detail, season,
interactions, and cache-and-logs, plus a shared.ts for the two
schemas reused across groups (serialGenreSchema, viewerInteractionSchema).
schemas.ts is now a barrel re-exporting all of them, so every existing
import path is unchanged. No schema content changed.
