+++ apps/web/src/features/serials/api/schemas.ts
@@ -1,400 +1,6 @@
-import { z } from "zod";
-import { personLinkSchema } from "@/features/people/shared";
-import { movieGenreSchema } from "@/types/api";
-
-export const tmdbSearchSeriesSchema = z.object({
-  id: z.number().int().positive(),
-  name: z.string(),
-  poster_path: z.string().nullable(),
-  first_air_date: z.string(),
-  overview: z.string(),
-});
-
-const serialGenreSchema = movieGenreSchema;
-
-const serialArchiveItemSchema = z.object({
-  tmdbId: z.number().int().positive(),
-  title: z.string(),
-  posterPath: z.string().nullable(),
-  backdropPath: z.string().nullable(),
-  firstAirDate: z.string().nullable(),
-  firstAirYear: z.number().int().nullable(),
-  creator: z.string().nullable(),
-  network: z.string().nullable(),
-  languageCode: z.string().nullable(),
-  genres: z.array(serialGenreSchema),
-  primaryGenre: z.string().nullable(),
-  logCount: z.number().int().nonnegative(),
-  avgRatingOutOfTen: z.number().nullable(),
-  tmdbRatingOutOfTen: z.number().nullable(),
-  ratedLogCount: z.number().int().nonnegative(),
-  numberOfEpisodes: z.number().int().nullable(),
-  viewerHasLogged: z.boolean(),
-  viewerWatchlisted: z.boolean(),
-  viewerFullyWatched: z.boolean(),
-  viewerHasProgress: z.boolean(),
-});
-
-const serialArchiveFeaturedSchema = z.object({
-  tmdbId: z.number().int().positive(),
-  title: z.string(),
-  posterPath: z.string().nullable(),
-  backdropPath: z.string().nullable(),
-  firstAirDate: z.string().nullable(),
-  firstAirYear: z.number().int().nullable(),
-  creator: z.string().nullable(),
-  network: z.string().nullable(),
-});
-
-const serialArchiveGenreCountSchema = z.object({
-  id: z.number().int().nullable().optional(),
-  name: z.string(),
-  count: z.number().int().nonnegative().nullable(),
-});
-
-export const serialArchiveSortSchema = z.enum([
-  "trending",
-  "first_air_desc",
-  "first_air_asc",
-  "logs_desc",
-  "rating_user_desc",
-  "rating_tmdb_desc",
-  "title_asc",
-]);
-
-export const serialArchivePeriodSchema = z.enum([
-  "all_time",
-  "this_year",
-  "last_10_years",
-  "this_week",
-  "today",
-]);
-
-export const serialArchiveResponseSchema = z.object({
-  totalCount: z.number().int().nonnegative(),
-  filteredCount: z.number().int().nonnegative(),
-  selectedGenre: z.string().nullable(),
-  selectedLanguage: z.string().nullable(),
-  selectedSort: serialArchiveSortSchema,
-  selectedPeriod: serialArchivePeriodSchema,
-  featuredSeries: serialArchiveFeaturedSchema.nullable(),
-  availableGenres: z.array(serialArchiveGenreCountSchema),
-  page: z.number().int().positive(),
-  limit: z.number().int().positive(),
-  hasMore: z.boolean(),
-  nextPage: z.number().int().positive().nullable(),
-  items: z.array(serialArchiveItemSchema),
-});
-
-const viewerInteractionSchema = z
-  .object({
-    watched: z.boolean(),
-    liked: z.boolean(),
-    rating: z.number().nullable(),
-    hasReview: z.boolean(),
-  })
-  .nullable();
-
-const serialDetailSeasonSchema = z.object({
-  id: z.number().int().positive(),
-  seasonNumber: z.number().int(),
-  name: z.string(),
-  episodeCount: z.number().int().nullable(),
-  airDate: z.string().nullable(),
-  posterPath: z.string().nullable(),
-  viewerInteraction: viewerInteractionSchema,
-});
-
-const serialDetailSeriesSchema = z.object({
-  id: z.number().int().positive(),
-  tmdbId: z.number().int().positive(),
-  title: z.string(),
-  originalTitle: z.string().nullable(),
-  posterPath: z.string().nullable(),
-  backdropPath: z.string().nullable(),
-  firstAirDate: z.string().nullable(),
-  firstAirYear: z.number().int().nullable(),
-  lastAirDate: z.string().nullable(),
-  creator: z.string().nullable(),
-  creators: z.array(personLinkSchema).default([]),
-  cast: z.array(personLinkSchema).default([]),
-  crew: z.array(personLinkSchema).default([]),
-  network: z.string().nullable(),
-  episodeRuntime: z.number().int().nullable(),
-  numberOfSeasons: z.number().int().nullable(),
-  numberOfEpisodes: z.number().int().nullable(),
-  status: z.string().nullable(),
-  overview: z.string().nullable(),
-  tagline: z.string().nullable(),
-  languageCode: z.string().nullable(),
-  genres: z.array(serialGenreSchema),
-  globalRating: z.number().nullable(),
-  globalRatingVoteCount: z.number().int().nullable(),
-  inProduction: z.boolean().nullable(),
-  seasons: z.array(serialDetailSeasonSchema),
-});
-
-export const serialDetailReviewSortSchema = z.enum(["popular", "recent"]);
-
-const serialDetailUserRatingSchema = z
-  .object({
-    diaryEntryId: z.string().nullable(),
-    reviewId: z.string().nullable(),
-    watchedDate: z.string().nullable(),
-    rewatch: z.boolean(),
-    rating: z.number().nullable(),
-    reviewContent: z.string().nullable(),
-    reviewContainsSpoilers: z.boolean().nullable(),
-  })
-  .nullable();
-
-const serialDetailReviewSchema = z.object({
-  id: z.string(),
-  content: z.string(),
-  containsSpoilers: z.boolean(),
-  createdAt: z.string(),
-  updatedAt: z.string(),
-  watchedDate: z.string().nullable(),
-  rating: z.number().nullable(),
-  likeCount: z.number().int().nonnegative(),
-  viewerHasLiked: z.boolean(),
-  author: z.object({
-    id: z.string(),
-    username: z.string(),
-    displayUsername: z.string().nullable(),
-    image: z.string().nullable(),
-    avatarUrl: z.string().nullable(),
-  }),
-  context: z
-    .object({
-      seasonNumber: z.number().int(),
-      episodeNumber: z.number().int().nullable(),
-      episodeName: z.string().nullable(),
-    })
-    .nullable(),
-});
-
-const serialDetailRatingBreakdownBucketSchema = z.object({
-  ratingValue: z.number().int().min(1).max(10),
-  count: z.number().int().nonnegative(),
-  percentage: z.number().int().min(0).max(100),
-});
-
-const similarSeriesItemSchema = z.object({
-  tmdbId: z.number().int(),
-  title: z.string(),
-  posterPath: z.string().nullable(),
-  firstAirYear: z.number().int().nullable(),
-});
-
-const serialDetailViewerTrackingSchema = z
-  .object({
-    watchedEpisodesCount: z.number().int().nonnegative(),
-    watchedEpisodes: z.array(
-      z.object({
-        seasonNumber: z.number().int(),
-        episodeNumber: z.number().int(),
-      }),
-    ),
-    currentEpisode: z
-      .object({
-        seasonNumber: z.number().int(),
-        episodeNumber: z.number().int(),
-        name: z.string(),
-      })
-      .nullable(),
-    ratingsCount: z.number().int().nonnegative(),
-    likesCount: z.number().int().nonnegative(),
-    reviewsCount: z.number().int().nonnegative(),
-  })
-  .nullable();
-
-export const serialDetailResponseSchema = z.object({
-  series: serialDetailSeriesSchema,
-  logsCount: z.number().int().nonnegative(),
-  reviewCount: z.number().int().nonnegative(),
-  userRating: serialDetailUserRatingSchema,
-  reviewsSort: serialDetailReviewSortSchema,
-  reviews: z.array(serialDetailReviewSchema),
-  ratingBreakdown: z.object({
-    totalRatedReviews: z.number().int().nonnegative(),
-    averageRating: z.number().nullable(),
-    buckets: z.array(serialDetailRatingBreakdownBucketSchema),
-  }),
-  similar: z.array(similarSeriesItemSchema),
-  viewerTracking: serialDetailViewerTrackingSchema,
-});
-
-const serialSeasonDetailEpisodeSchema = z.object({
-  id: z.number().int().positive(),
-  seasonNumber: z.number().int().nonnegative(),
-  episodeNumber: z.number().int().positive(),
-  name: z.string(),
-  overview: z.string().nullable(),
-  airDate: z.string().nullable(),
-  stillPath: z.string().nullable(),
-  runtimeMinutes: z.number().int().nullable(),
-  runtimeLabel: z.string().nullable(),
-  viewerInteraction: viewerInteractionSchema,
-});
-
-export const serialSeasonDetailSchema = z.object({
-  tmdbId: z.number().int().positive(),
-  season: z.object({
-    id: z.number().int().positive(),
-    seasonNumber: z.number().int().nonnegative(),
-    name: z.string(),
-    overview: z.string().nullable(),
-    airDate: z.string().nullable(),
-    posterPath: z.string().nullable(),
-    episodeCount: z.number().int().nonnegative(),
-  }),
-  episodes: z.array(serialSeasonDetailEpisodeSchema),
-});
-
-export const serialInteractionSchema = z
-  .object({
-    liked: z.boolean(),
-    watchlisted: z.boolean(),
-    rating: z.number().min(0.5).max(10).multipleOf(0.5).nullable(),
-    watched: z.boolean(),
-  })
-  .passthrough();
-
-export const updateSerialInteractionInputSchema = z
-  .object({
-    liked: z.boolean().optional(),
-    watchlisted: z.boolean().optional(),
-    rating: z.number().min(0.5).max(10).multipleOf(0.5).nullable().optional(),
-    watched: z.boolean().optional(),
-  })
-  .refine(
-    (payload) =>
-      payload.liked !== undefined ||
-      payload.watchlisted !== undefined ||
-      payload.rating !== undefined ||
-      payload.watched !== undefined,
-    {
-      message: "At least one interaction field is required",
-    },
-  );
-
-const serialLogRatingSchema = z.number().min(0.5).max(10).multipleOf(0.5);
-
-export const createSeriesLogInputSchema = z.object({
-  watchedDate: z.string(),
-  rating: serialLogRatingSchema.optional(),
-  rewatch: z.boolean().optional(),
-  review: z.string().max(5000).optional(),
-  containsSpoilers: z.boolean().optional(),
-});
-
-export const createSeriesLogResponseSchema = z.object({
-  entry: z
-    .object({
-      id: z.string(),
-      userId: z.string(),
-      seriesId: z.number().int(),
-      watchedDate: z.string(),
-      rating: z.number().nullable(),
-      rewatch: z.boolean(),
-      createdAt: z.string(),
-      updatedAt: z.string(),
-    })
-    .passthrough(),
-  series: z
-    .object({
-      id: z.number().int(),
-      tmdbId: z.number().int(),
-      title: z.string(),
-      posterPath: z.string().nullable(),
-      firstAirYear: z.number().int().nullable(),
-    })
-    .passthrough(),
-  review: z
-    .object({
-      id: z.string(),
-      content: z.string(),
-      containsSpoilers: z.boolean(),
-    })
-    .nullable(),
-});
-
-export const cachedSeriesSchema = z
-  .object({
-    id: z.number().int().positive(),
-    tmdbId: z.number().int().positive(),
-    title: z.string(),
-    originalTitle: z.string().nullable(),
-    posterPath: z.string().nullable(),
-    backdropPath: z.string().nullable(),
-    firstAirDate: z.string().nullable(),
-    firstAirYear: z.number().int().nullable(),
-    lastAirDate: z.string().nullable(),
-    creator: z.string().nullable(),
-    network: z.string().nullable(),
-    episodeRuntime: z.number().int().nullable(),
-    numberOfSeasons: z.number().int().nullable(),
-    numberOfEpisodes: z.number().int().nullable(),
-    status: z.string().nullable(),
-    overview: z.string().nullable(),
-    tagline: z.string().nullable(),
-    languageCode: z.string().nullable(),
-    genres: z.array(serialGenreSchema).nullish(),
-    cachedAt: z.string(),
-  })
-  .passthrough();
-
-export const trendingSeriesSchema = z.object({
-  tmdbId: z.number().int().positive(),
-  title: z.string(),
-  posterPath: z.string().nullable(),
-  firstAirYear: z.number().int().nullable(),
-});
-
-export const trendingSeriesListSchema = z.array(trendingSeriesSchema);
-export const tmdbSearchSeriesListSchema = z.array(tmdbSearchSeriesSchema);
-
-export const serialDiaryEntrySchema = z.object({
-  id: z.string(),
-  watchedDate: z.string(),
-  rating: z.number().nullable(),
-  rewatch: z.boolean(),
-  seriesId: z.number().int(),
-  createdAt: z.string(),
-  updatedAt: z.string(),
-  seriesTmdbId: z.number().int(),
-  seriesTitle: z.string(),
-  seriesPosterPath: z.string().nullable(),
-  seriesFirstAirYear: z.number().int().nullable(),
-  reviewId: z.string().nullable(),
-  reviewContent: z.string().nullable(),
-  reviewContainsSpoilers: z.boolean().nullable(),
-  reviewCreatedAt: z.string().nullable(),
-});
-
-export const serialDiaryListSchema = z.array(serialDiaryEntrySchema);
-
-export const updateSerialLogInputSchema = z.object({
-  watchedDate: z.string().optional(),
-  rating: z.number().min(0.5).max(10).multipleOf(0.5).nullable().optional(),
-  rewatch: z.boolean().optional(),
-});
-
-export const serialLogSchema = z.object({
-  diaryEntryId: z.string(),
-  watchedDate: z.string(),
-  rating: z.number().nullable(),
-  rewatch: z.boolean(),
-  createdAt: z.string(),
-  username: z.string(),
-  userDisplayName: z.string().nullable(),
-  avatarUrl: z.string().nullable(),
-  reviewContent: z.string().nullable(),
-  reviewContainsSpoilers: z.boolean().nullable(),
-  reviewUpdatedAt: z.string().nullable(),
-});
-
-export const serialLogsListSchema = z.array(serialLogSchema);
-
-export { serialArchiveItemSchema };
+export * from "./schemas/search";
+export * from "./schemas/archive";
+export * from "./schemas/detail";
+export * from "./schemas/season";
+export * from "./schemas/interactions";
+export * from "./schemas/cache-and-logs";
