+++ apps/web/src/features/serials/api/schemas/shared.ts
@@ -1,0 +1,13 @@
+import { z } from "zod";
+import { movieGenreSchema } from "@/types/api";
+
+export const serialGenreSchema = movieGenreSchema;
+
+export const viewerInteractionSchema = z
+  .object({
+    watched: z.boolean(),
+    liked: z.boolean(),
+    rating: z.number().nullable(),
+    hasReview: z.boolean(),
+  })
+  .nullable();
